Metric timeline (#440)
* Move the timeline to the common code

* Use the timeline for the text metric too

* fix the height

* make linking optional

* Add the selection widget

* fix incorrect nullable

* clean up

* more

* simplify tolocal

* more clean up

* start work on better touch target for dots event

* split the timeline

* Add a touch target for small metric

* add a state to the async to fix some refresh issues

* fix the widget

* add cards

* Add an historgram for the text metric

* improve the text

* add touch on the histogram

* extract logic from the graph

* use more common

* more fix

* fix range

+++ flutter_frontend/lib/ui/blocs/metrics/detail/stats_widgets/metric_histogram.dart
@@ -30,7 +30,7 @@ class MetricHistogram extends StatelessWidget {
     for (final step in steps.entries) {
       bars.add(
         BarChartGroupData(
-          x: step.key.toInt(),
+          x: step.key,
           barRods: [
             BarChartRodData(
               toY: step.value,
